backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
bcrypt<4.0.0